Remove Info button from header to give a more uniform appearance
In addition I also removed the max-height rule from the header component to allow for a seamless transition to the multitabpane no matter how wide the viewing screen is.

As it stood, I was trying to use padding to always make the countdown timer touch the multitab pane as the header could be too tall for it. Now that is not an issue.