Functionality added to interest form in the sense that there are react states keeping track of the value of each input and a submit button that can be used to launch a POST connection to a mongodb backend.
An idea for the future is to replace the "plus-one" radio question with a number input that is a "How Many People in your Party?" question. This can also tie into a react conditional rendering where a text input will be rendered for however many quests are in the party, allowing us to gather all attendee names.