Frontend overhaul. The Interest Form has been nixxed in favor of going straight to an rsvp system. Initial form built reusing radio question component from Interest Form
The form also dynamically renders the correct number of textfields needed for how many guests the user specifies, and I added CSS breakpoints to render the radio answers vertically when the screen gets too narrow.

Next commit will be to duplicate the model and controller in the backend and tweak them for use with the rsvp system.