Added a useState to track all guestNames being written in the rsvp form.
It was trickier than expected as I had not realized useState triggeres a re-render each time the setState function is called, inadvertently creating an infinite loop while trying to dynamically render textfields and set the proper amount of blank spaces in the guestName state.

This was remedied by using a useEffect that listens to the guestCount state, this way each time the guestCount is updated the guestName list is dynamically rendered in the useEffect a single time and the textfields can render the correct amount of times.