move as much lgoic out of the rsvp jsx file and move it into it's own module in the js file. Rsvp.js is as minimized as I will make it now.
I'm sure I could shrink it more, but to fixate on that would be to not move onto the next step of the project which it is time to do.

Next steps are to finish the onSubmit function, having it organize the information stored in states and then add it to the mongoose database for later retreival.