RSVP submission finished being built out. RSVP tab hidden until it is ready for public deployment.
FAQ section added publicly.

At this point, the backend branch is ready to be merged back into the main branch once the faq tab is ready to be deployed. It technically should have been it's own branch but is lumped into backend to keep the number of branching branches down and hopefully make the merge back to main les of a headache as there are commits on the main that are not in backend